website_django/requirements.txt
Django==4.2.27
gunicorn==21.2.0
oracledb==2.5.*
python-docx==1.1.2
requests==2.32.3
